Set leagcy grub as default for xen pv guest
Since the deployed Xen hosts currently do not know about the new grub2
config files, installing 12.2 and Factory in a Xen PV guest will fail.

To give a good user impression for 12.2 installations withing a guest,
continue to use grub as a bootloader if a Xen PV guest is detected.