use mbr if /boot is btrfs and on logical partition
For btrfs, it's not possible to use block-list installation and
would require embedding core.img in btrfs's bootloader area or
in mbr and mbr hole which could provide enough room for it.

If btrfs /boot is on logical partition, we can't propose to use
extended paritition as it's too small in size (1024 bytes) to
the core.img and can't propose logical partition as it cannot
be chain loaded by generic mbr boot code. The only remaining
choice is mbr.