Make GRUB_DISTRIBUTOR with product version
First the Product::name returns version on running systems but not on
the installation system. We have to ensure that the version is specified
to GRUB_DISTRIBUTOR anyway.

Without the version, EFI boot menu would be added as "openSUSE" for all
versions and reference same bootloader path. This casues problem on
multiple version installation.

This fixes EFI boot menu entry cannot be chosen and multiple openSUSE
installations will overwrite it (bnc#796514).