Fix LVM installation warning and location
As grub2 supports booting kernel on a LVM partition, remove the warning
message that is obseleted. Also the install location is refined to
restrict on mbr if the partition boot record is not allowed to install,
like in this case or dm raid as the partition boot record is used as
data or meta data blocks and blocklist addressing won't work either.