get rid of grub suggestion
We ask the probing agent for the loader type. Unfortunately this still
returns grub on x86. I sugggest to skip this step altogether. We decide the
loader type in yast2-bootloader anyway. If we want to enable it on openSUSE
we should do it in this function somewhere.

Not sure where grub2 on ppc64le comes from, though...