Explain why write stage1 to several devices
Before this whole bug fix, the code remembered the "boot partition" and
installed stage 1 to one of its underlying devices.

Now, it remembers the "boot mount point" (which is / or /boot*) and uses
 #stage1_disks_for or #stage1_partitions_for (according to a UI choice)
to get ALL underlying devices for it (as an Array) so that the (Legacy)
BIOS does not have a chance to pick an empty BR to boot from.

*) more correctly, the Filesystem corresponding to that mount point