Support powerpc secure boot (bsc#1192764 jsc#SLE-18271).
This adds powerpc secure boot detection support to systeminfo and grub
install.

Currently we have secure boot support indicated by ibm,secure-boot file
presence but in many configurations this file is present but secure boot
cannot work.

We should propose secure boot on powerpc when and only when content of
this file is non-zero because then booting without secure boot will
fail.

On the other hand we should enable the user to manually turn on secure
boot when ibm,secure-boot is present or even in all cases.

In the future it is expected that the firmware grows something like
efivarfs allowing the OS to pass additional parameters to the secure
bootloader which should make the secure boot reliable.

Signed-off-by: Michal Suchanek <msuchanek@suse.de>